+++ src/api/runners/mod.rs
@@ -5,5 +5,6 @@
 //! backend turn runners themselves) move here so `mission_runner.rs` can
 //! shrink down to orchestration (dispatch, retry/fallback, TerminalReason).
 
+pub(crate) mod codex;
 pub(crate) mod errors;
 pub(crate) mod grok;
